requirements.txt
fastapi[standard]
pydantic[email]
pydantic-settings
python-dotenv
sqlalchemy
alembic
psycopg2-binary
httpx
